requirements.txt
pandas
numpy
matplotlib
tensorflow
opencv-python
seaborn
scikit-learn
pillow
dask
dask[complete]
streamlit
#-e .